backend/requirements.txt
# Django and REST Framework
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
djoser==2.2.2

# Database and History
psycopg2-binary==2.9.10
django-simple-history==3.4.0

# Environment and Configuration
python-dotenv==1.0.0
python-decouple==3.8

# WSGI server for production
gunicorn==21.2.0

# Filtering and API features
django-filter==23.5
django-reversion==5.0.8

# Date/Time handling
pytz==2023.3

# Production dependencies
whitenoise==6.6.0  # For serving static files
dj-database-url==2.1.0  # For database URL configuration

#recipe srapers
recipe-scrapers==15.4.0

# New packages
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Scanner dependencies
psutil>=5.9.0  # For process monitoring
playwright>=1.40.0  # For web scraping
sendgrid>=6.10.0  # For email notifications
twilio>=8.2.0  # For SMS notifications (optional)
geopy>=2.3.0  # For distance calculations
# AI/LLM dependencies
httpx>=0.24.0  # Async HTTP client for image fetching
google-genai>=0.2.0  # Google Gemini API (new package)
